docs: added TSDoc annotations
