+++ libs/rxidb/src/lib/object-store/get-object-store.ts
@@ -1,7 +1,40 @@
-import { upgradeDatabase } from '../database';
+import { upgradeDatabase } from '../database/upgrade-database';
 import { filter, Observable, of, ReplaySubject, share, Subject, switchMap, takeUntil } from 'rxjs';
 import { DATABASE_DELETE_EVENTS } from '../rxidb-internal.events';
 
+/**
+ * Converts an Observable<IDBDatabase> stream to an Observable<IDBObjectStore> stream
+ *
+ * This operator checks for existing ObjectStore instances with the provided store name.
+ * If there is no previous instance it creates the ObjectStore, otherwise creates a `readwrite`
+ * transaction
+ *
+ * @example
+ * const database$ = connectIndexedDb('test_db');
+ * // creates an ObjectStore with key-value pairs
+ * const keyValueStore$ = database$.pipe(getObjectStore('test_store'))
+ * // creates an ObjectStore with autoIncrement
+ * const autoIncrementStore$ = database$.pipe(getObjectStore('test_autoincrement_store', { autoIncrement: true }))
+ *
+ * // subscribing to the observable will create the database and the store
+ * keyValueStore$.subscribe({
+ *   next: (db: IDBObjectStore) => console.log(store.name),
+ *   error: (e) => console.error(e),
+ *   complete: () => console.warn('database deleted')
+ * });
+ *
+ * autoIncrementStore$.subscribe({
+ *   next: (db: IDBObjectStore) => console.log(store.name),
+ *   error: (e) => console.error(e),
+ *   complete: () => console.warn('database deleted')
+ * })
+ *
+ * @remarks Completes when the source database gets deleted.
+ * @param name
+ * @param options - IDBObjectStoreParameters
+ *
+ * @returns Observable<IDBObjectStore>
+ */
 export function getObjectStore(
   name: string,
   options?: IDBObjectStoreParameters
@@ -9,13 +42,13 @@ export function getObjectStore(
   return (s$) =>
     s$.pipe(
       switchMap((existingDb: IDBDatabase) => {
-        const isExistingDatabase = existingDb.objectStoreNames.contains(name);
-        const upgrade = isExistingDatabase ? of(existingDb) : upgradeDatabase(existingDb);
+        const isExistingObjectStore = existingDb.objectStoreNames.contains(name);
+        const upgrade = isExistingObjectStore ? of(existingDb) : upgradeDatabase(existingDb);
         return upgrade.pipe(
           switchMap((db: IDBDatabase) => {
             const storeSubject = new Subject<IDBObjectStore>();
             let store: IDBObjectStore;
-            if (isExistingDatabase) {
+            if (isExistingObjectStore) {
               store = db.transaction(name, 'readwrite').objectStore(name);
             } else {
               store = options ? db.createObjectStore(name, options) : db.createObjectStore(name);
